Title: 세 수의 합과 평균 구하기, Time: 156ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240619/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -3,13 +3,13 @@ import java.util.Scanner;
 public class Main {
     public static void main (String args[]) {
         Scanner sc = new Scanner(System.in);
-        double a = sc.nextInt();
-        double b = sc.nextInt();
-        double c = sc.nextInt();
+        int a = sc.nextInt();
+        int b = sc.nextInt();
+        int c = sc.nextInt();
     
 
-        System.out.printf("%.0f\n",a+b+c);
-        System.out.printf("%.0f",(a+b+c)/3);
+        System.out.printf("%d\n",a+b+c);
+        System.out.printf("%d",(a+b+c)/3);
 
     }
 }
